+++ src/contents/ui/Day.qml
@@ -11,6 +11,8 @@ import QtQuick 2.1
 import QtQuick.Layouts 1.1
 
 Rectangle {
+	id: container
+
 	property string label: '?'
 	property string fg: "FFffffff"
 	property string bg: "#00000000"
@@ -24,7 +26,7 @@ Rectangle {
 	Layout.minimumWidth: weekGrid.cellMinWidth
 	Layout.alignment: Qt.AlignVCenter | Qt.AlignHCenter
 	Layout.margins: 0
-	
+
 	color: bg
 
 	Text {
@@ -35,7 +37,7 @@ Rectangle {
 		font.bold: bold
 		font.italic: italic
 
-		anchors.fill: parent
+		anchors.fill: container
 
 		Layout.fillWidth: true
 		Layout.margins: 4